created insert.sql for test purposes and changed min and max price to decimals

+++ Code/src/init_data/create.sql
@@ -13,7 +13,7 @@ CREATE TABLE jobs(
     job_id INTEGER PRIMARY KEY,
     description VARCHAR(200),
     requester CHAR,
-    minPrice INTEGER,
-    maxPrice INTEGER,
+    minPrice DECIMAL,
+    maxPrice DECIMAL,
     FOREIGN KEY (requester) REFERENCES users(email)
 );
